+++ test/unit/server/services/procedures/google-maps.spec.js
@@ -29,7 +29,9 @@ describe(utils.suiteName(__filename), function() {
             ['maxLatitude'],
             ['minLongitude'],
             ['getDistance', ['startLatitude', 'startLongitude', 'endLatitude', 'endLongitude']],
-            ['minLatitude']
+            ['minLatitude'],
+            ['getRadarTimeOffsets'],
+            ['getRadarMap', ['latitude', 'longitude', 'width', 'height', 'zoom', 'timeOffset', 'options']],
         ]);
     });
